Add files via upload
The Russian localization of the "Not Enough Time" game as part of the LocJAM 6 contest. LocJam is an online game translation marathon focused on creativity, discovery and collaboration. The game to translate was a web-based point-and-click adventure with 3500 words in text format with control codes (.lua).

Here you can test my Russian localization or even play this game in its original language (English): 

https://der-artur.itch.io/not-enough-time-ru-loc-by-der-artur